backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.8.2
pydantic-settings==2.4.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.1
python-dotenv==1.0.1
spacy==3.7.4
sentence-transformers==3.1.1
faiss-cpu==1.8.0.post1
langgraph==0.0.55
autogen==0.7.0
httpx==0.27.2
